requirements.txt
google-api-python-client
openai
python-dotenv